+++ tests/test_processing/test_validation_manager/test_unescape_string.py
@@ -1,15 +1,15 @@
 import pytest
-from mite_extras.processing.validation_manager import ValidationManager
+from mite_extras.processing.validation_manager import MoleculeValidator
 
 
 @pytest.fixture
-def validation_manager():
-    return ValidationManager()
+def _molecule_validator():
+    return MoleculeValidator()
 
 
-def test_unescape_string(validation_manager):
+def test_unescape_string(_molecule_validator):
     """Test canonicalizing a valid SMARTS string"""
     string = r"This\\is\a\\test\\string\\with\\double\backslashes"
-    unescaped_string = validation_manager.unescape_string(string)
+    unescaped_string = _molecule_validator._clean_string(string)
     truth = r"This\is\a\test\string\with\double\backslashes"
     assert unescaped_string == truth