+++ tests/test_processing/test_validation_manager/test_validate_reaction_smarts.py
@@ -1,41 +1,41 @@
 import pytest
-from mite_extras.processing.validation_manager import ValidationManager
-from rdkit.Chem import CanonSmiles, MolFromSmiles, MolToSmiles
-from rdkit.Chem.rdChemReactions import ReactionFromSmarts
+from mite_extras.processing.validation_manager import (
+    ReactionValidator,
+)
 
 
 @pytest.fixture
-def validation_manager():
-    return ValidationManager()
+def _reaction_validator():
+    return ReactionValidator()
 
 
-def test_validate_reaction_smarts_valid(validation_manager):
+def test_validate_reaction_valid(_reaction_validator):
     """Test validating a reaction SMARTS string with valid expected and forbidden products"""
     reaction_smarts = "[C:1][O:2]>>[C:1]=[O:2]"
     substrate_smiles = "CO"
     expected_products = ["C=O"]
     forbidden_products = []
-    result = validation_manager.validate_reaction_smarts(
+    result = _reaction_validator.validate_reaction(
         reaction_smarts, substrate_smiles, expected_products, forbidden_products
     )
     assert result == None
 
 
-def test_validate_reaction_smarts_forbidden_in_expected(validation_manager):
+def test_validate_reaction_forbidden_in_expected(_reaction_validator):
     """Test when forbidden products are mistakenly included in expected products"""
     reaction_smarts = "[C:1][O:2]>>[C:1]=[O:2]"
     substrate_smiles = "CO"
     expected_products = ["C=O"]
     forbidden_products = ["C=O"]  # Forbidden product is also in expected products
     with pytest.raises(
-        ValueError, match="Some expected products are listed as forbidden products."
+        ValueError, match="Some expected products are listed as forbidden"
     ):
-        validation_manager.validate_reaction_smarts(
+        _reaction_validator.validate_reaction(
             reaction_smarts, substrate_smiles, expected_products, forbidden_products
         )
 
 
-def test_validate_reaction_smarts_unexpected_products(validation_manager):
+def test_validate_reaction_unexpected_products(_reaction_validator):
     """Test when the predicted products do not match expected products"""
     reaction_smarts = "[C:1][O:2]>>[C:1]=[O:2]"
     substrate_smiles = "OCO"
@@ -44,81 +44,77 @@ def test_validate_reaction_smarts_unexpected_products(validation_manager):
     predicted_smiles_set = {"O=CO"}  # Assume this is what the reaction predicts
     with pytest.raises(
         ValueError,
-        match=f"Products '{predicted_smiles_set}' do not meet expectations '{set(expected_products)}'.",
+        match=f"Products {predicted_smiles_set} do not include all expected products {set(expected_products)}",
     ):
-        validation_manager.validate_reaction_smarts(
+        _reaction_validator.validate_reaction(
             reaction_smarts, substrate_smiles, expected_products, forbidden_products
         )
 
 
-def test_validate_reaction_smarts_forbidden_products(validation_manager):
-    """Test when forbidden products are found in the reaction output"""
+def test_validate_reaction_forbidden_products(_reaction_validator):
+    """Test when forbidden products are found in reaction output"""
     reaction_smarts = "[C:1][O:2]>>[C:1]=[O:2]"
     substrate_smiles = "OCO |LN:1:1.2|"
     expected_products = ["O=CO"]
     forbidden_products = ["O=CCO"]  # Forbidden product is also expected
     with pytest.raises(
-        ValueError, match="Forbidden products were found in the reaction output."
+        ValueError, match="Forbidden products were found in reaction output"
     ):
-        validation_manager.validate_reaction_smarts(
+        _reaction_validator.validate_reaction(
             reaction_smarts, substrate_smiles, expected_products, forbidden_products
         )
 
 
-def test_validate_reaction_smarts_empty_expected_products(validation_manager):
+def test_validate_reaction_empty_expected_products(_reaction_validator):
     """Test when the expected products list is empty"""
     reaction_smarts = "[C:1][O:2]>>[C:1]=[O:2]"
     substrate_smiles = "CO"
     expected_products = []  # Empty expected products list
     forbidden_products = []
-    with pytest.raises(ValueError, match="Expected products list cannot be empty."):
-        validation_manager.validate_reaction_smarts(
+    with pytest.raises(ValueError, match="Expected products list cannot be empty"):
+        _reaction_validator.validate_reaction(
             reaction_smarts, substrate_smiles, expected_products, forbidden_products
         )
 
 
-def test_validate_reaction_smarts_invalid_expected_product(validation_manager):
+def test_validate_reaction_invalid_expected_product(_reaction_validator):
     """Test when one of the expected products cannot be parsed"""
     reaction_smarts = "[C:1][O:2]>>[C:1]=[O:2]"
     substrate_smiles = "CO"
     expected_products = ["C=O", "INVALID_SMILES"]  # Invalid product
     forbidden_products = []
-    with pytest.raises(
-        ValueError, match=f"Could not read SMILES string 'INVALID_SMILES'"
-    ):
-        validation_manager.validate_reaction_smarts(
+    with pytest.raises(ValueError, match=f"Invalid SMILES string: 'INVALID_SMILES'"):
+        _reaction_validator.validate_reaction(
             reaction_smarts, substrate_smiles, expected_products, forbidden_products
         )
 
 
-def test_validate_reaction_smarts_invalid_forbidden_product(validation_manager):
+def test_validate_reaction_invalid_forbidden_product(_reaction_validator):
     """Test when one of the forbidden products cannot be parsed"""
     reaction_smarts = "[C:1][O:2]>>[C:1]=[O:2]"
     substrate_smiles = "CO"
     expected_products = ["C=O"]
     forbidden_products = ["INVALID_SMILES"]  # Invalid forbidden product
-    with pytest.raises(
-        ValueError, match="Could not read SMILES string 'INVALID_SMILES'"
-    ):
-        validation_manager.validate_reaction_smarts(
+    with pytest.raises(ValueError, match="Invalid SMILES string: 'INVALID_SMILES'"):
+        _reaction_validator.validate_reaction(
             reaction_smarts, substrate_smiles, expected_products, forbidden_products
         )
 
 
-def test_validate_reaction_smarts_valid_composite_pass(validation_manager):
+def test_validate_reaction_valid_composite_pass(_reaction_validator):
     """Test when composite SMILES (Rhea)"""
     # TODO (AR 2024-08-15): @MMZ not sure the reaction gives two products on this one
     reaction_smarts = "[cH:9]1[cH:8][c:7]2[cH:6][cH:5][cH:4][cH:3][c:2]2[nH:1]1.[ClH:10]>>[Cl:10][c:9]1[cH:8][cH:7][cH:6][c:5]2[nH:4][cH:3][cH:2][c:1]12"
     substrate_smiles = "c1cc2ccccc2[nH]1.Cl"
     expected_products = ["Clc1cccc2[nH]ccc12"]
     forbidden_products = []
-    result = validation_manager.validate_reaction_smarts(
+    result = _reaction_validator.validate_reaction(
         reaction_smarts, substrate_smiles, expected_products, forbidden_products
     )
     assert result == None
 
 
-def test_validate_reaction_smarts_frequency_variation(validation_manager):
+def test_validate_reaction_frequency_variation(_reaction_validator):
     """Test when position variation"""
     reaction_smarts = "[NH2:1][C@@H:2]([CH2:3][c:4]1[cH:5][n:6][c:7]2[cH:8][cH:9][cH:10][cH:11][c:12]12)[C:13]([OH:14])=[O:15].[ClH:16]>>[NH2:1][C@@H:2]([CH2:3][c:4]1[cH:5][nH:6][c:7]2[cH:8][cH:9][cH:10][c:11]([Cl:12])[c:13]12)[C:14]([OH:15])=[O:16] |r,Sg:n:2:1-2:ht|"  # Example SMARTS with frequency variation
     substrate_smiles = "N[C@@H](Cc1c[nH]c2ccccc12)C(O)=O.Cl"
@@ -127,13 +123,13 @@ def test_validate_reaction_smarts_frequency_variation(validation_manager):
         "N[C@@H](CCc1c[nH]c2cccc(Cl)c12)C(O)=O",
     ]
     forbidden_products = []
-    result = validation_manager.validate_reaction_smarts(
+    result = _reaction_validator.validate_reaction(
         reaction_smarts, substrate_smiles, expected_products, forbidden_products
     )
     assert result == None
 
 
-def test_validate_reaction_smarts_frequency_variation_unordered(validation_manager):
+def test_validate_reaction_frequency_variation_unordered(_reaction_validator):
     """Test when position variation"""
     reaction_smarts = "[NH2:1][C@@H:2]([CH2:3][c:4]1[cH:5][n:6][c:7]2[cH:8][cH:9][cH:10][cH:11][c:12]12)[C:13]([OH:14])=[O:15].[ClH:16]>>[NH2:1][C@@H:2]([CH2:3][c:4]1[cH:5][nH:6][c:7]2[cH:8][cH:9][cH:10][c:11]([Cl:12])[c:13]12)[C:14]([OH:15])=[O:16] |r,Sg:n:2:1-2:ht|"  # Example SMARTS with frequency variation
     substrate_smiles = "Cl.N[C@@H](Cc1c[nH]c2ccccc12)C(O)=O"
@@ -142,13 +138,13 @@ def test_validate_reaction_smarts_frequency_variation_unordered(validation_manag
         "N[C@@H](CCc1c[nH]c2cccc(Cl)c12)C(O)=O",
     ]
     forbidden_products = []
-    result = validation_manager.validate_reaction_smarts(
+    result = _reaction_validator.validate_reaction(
         reaction_smarts, substrate_smiles, expected_products, forbidden_products
     )
     assert result == None
 
 
-def test_validate_reaction_smarts_halogenation(validation_manager):
+def test_validate_reaction_halogenation(_reaction_validator):
     """Test when multiple atoms allowed (halogenation)"""
     reaction_smarts = "[#7:1]-[#6:2](-[#6:3]-[c:4]1[c:5][nH1:6][c:7]2[c:8][c:9][c:10][c:11][c:12]12)-[#6:13](-[#8:14])=[O:15]>>[#7:1]-[#6:2](-[#6:3]-[c:4]1[c:5][nH1:6][c:7]2[c:8][c:9](-[#17,#35:10])[c:11][c:12][c:13]12)-[#6:14](-[#8:15])=[O:16]"  # Example SMARTS with Cl OR Br
     substrate_smiles = "NC(Cc1c[nH]c2ccccc12)C(=O)O"
@@ -157,7 +153,26 @@ def test_validate_reaction_smarts_halogenation(validation_manager):
         "NC(Cc1c[nH]c2cc(Cl)ccc12)C(=O)O",
     ]
     forbidden_products = []
-    result = validation_manager.validate_reaction_smarts(
+    result = _reaction_validator.validate_reaction(
         reaction_smarts, substrate_smiles, expected_products, forbidden_products
+    )
+    assert result == None
+
+
+def test_validate_intramolecular_macrolactam(_reaction_validator):
+    reaction_smarts = "([#7&H2:1]-[#6&H2:2]-[#6:3](-[#7&H1:5])=[#8:4].[#7&H1:6]-[#6:7](-[#6:13](-[#7&H1:15])=[#8:14])-[#6:8]-[#6:9]-[#6:10](-[#8&H1:12])=[#8:11])>>[#7:1](-[#8:12]-[#6:10](=[#8:11])-[#6:9]-[#6:8]-[#6:7](-[#6:13](-[#7:15])=[#8:14])-[#7:6])-[#6:2]-[#6:3](-[#7:5])=[#8:4]"
+    substrate_smiles = (
+        "C[C@H](NC(=O)[C@H](CCC(=O)O)NC(=O)[C@H](C)NC(=O)[C@H](C)NC(=O)CN)C(=O)O"
+    )
+    expected_products = [
+        "C[C@H](NC(=O)[C@@H]1CCC(=O)ONCC(=O)N[C@@H](C)C(=O)N[C@@H](C)C(=O)N1)C(=O)O"
+    ]
+    forbidden_products = []
+    result = _reaction_validator.validate_reaction(
+        reaction_smarts,
+        substrate_smiles,
+        expected_products,
+        forbidden_products,
+        intramolecular=True,
     )
     assert result == None