+++ CHANGELOG.md
@@ -5,6 +5,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 This project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.4.0] 02-12-2024
+
+### Changed
+
+- Bumped `mite_schema` version to `1.7.0`
+- Reworked `ValidationManager`
+- Implemented option for intramolecular reactions (recognized from reaction SMARTS)
+
 ## [1.3.1] 02-12-2024
 
 ### Added